moved component and fixed bug with observables
