+++ src/interface/src/app/data-layers/data-layer-name/data-layer-name.component.html
@@ -1,0 +1,7 @@
+<button
+  class="data-layer-name"
+  (click)="goToSelectedLayer(selectedDataLayer)"
+  *ngIf="selectedDataLayer$ | async as selectedDataLayer">
+  <mat-icon class="icon-back">arrow_back</mat-icon>
+  {{ selectedDataLayer.name }}
+</button>
