+++ src/interface/src/app/data-layers/data-layer-name/data-layer-name.component.spec.ts
@@ -2,7 +2,7 @@ import { ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { DataLayerNameComponent } from './data-layer-name.component';
 import { MockProvider } from 'ng-mocks';
-import { DataLayersStateService } from '../../data-layers/data-layers.state.service';
+import { DataLayersStateService } from '../data-layers.state.service';
 import { of } from 'rxjs';
 
 describe('DataLayerNameComponent', () => {
