--- src/interface/src/app/maplibre-map/data-layer-name/data-layer-name.component.html
@@ -1,10 +1,0 @@
-<mgl-control
-  position="bottom-left"
-  *ngIf="selectedDataLayer$ | async as selectedDataLayer">
-  <button
-    class="data-layer-name"
-    (click)="goToSelectedLayer(selectedDataLayer)">
-    <mat-icon class="icon-back">arrow_back</mat-icon>
-    {{ selectedDataLayer.name }}
-  </button>
-</mgl-control>
