+++ src/interface/src/app/treatments/treatment-map/treatment-map.component.html
@@ -72,5 +72,7 @@
     click to apply treatment on single stand or drag to apply on multiple stands
   </app-map-tooltip>
 
-  <app-data-layer-name></app-data-layer-name>
+  <mgl-control position="bottom-left">
+    <app-data-layer-name></app-data-layer-name>
+  </mgl-control>
 </mgl-map>
